Revert "Remove `HasCorporateInformationPages`"
This reverts commit f8ccaf74786936f5190f7215d8b482938b338a6a.
